fix my info to tab components
